Create travelcompanion_flask.py
Created a flask environment on python to provide inference services especially on the network. We imported flask_ngrok so that it is possible to work through Google Colab. The file also imported the model1.h5 and weight1.h5 as well as tokenizer.json to use the same formatting according to the previous training. This python will in turn provide a request and post API.